backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7
reportlab==4.0.4
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.15.0
